Adjust copy_dir() for train.py and notebook
