+++ src/models/CNNDepthMap/CNNDepthMap-height/q3-depthmap-plaincnn-height/src/train.py
@@ -16,25 +16,9 @@ from wandb.keras import WandbCallback
 from config import CONFIG
 from constants import BLACKLIST_QRCODES, MODEL_CKPT_FILENAME, REPO_DIR
 from model import create_cnn
+from train_util import copy_dir
 
 logging.basicConfig(level=logging.INFO, format='%(asctime)s - %(levelname)s - %(message)s - %(pathname)s: line %(lineno)d')
-
-
-def copy_dir(src: Path, tgt: Path, glob_pattern: str, should_touch_init: bool = False):
-    logging.info("Creating temp folder")
-    if tgt.exists():
-        shutil.rmtree(tgt)
-    tgt.mkdir(parents=True, exist_ok=True)
-    if should_touch_init:
-        (tgt / '__init__.py').touch(exist_ok=False)
-
-    paths_to_copy = list(src.glob(glob_pattern))
-    logging.info(f"Copying to {tgt} the following files: {str(paths_to_copy)}")
-    for p in paths_to_copy:
-        destpath = tgt / p.relative_to(src)
-        destpath.parent.mkdir(parents=True, exist_ok=True)
-        shutil.copy(p, destpath)
-
 
 # Get the current run.
 run = Run.get_context()
